% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebio_sra.R
\name{ebio_sra_extract}
\alias{ebio_sra_extract}
\title{Generate a table from MPIBT SRA file structure}
\usage{
ebio_sra_extract(directory)
}
\arguments{
\item{directory}{Project directory to scan.}
}
\value{
Dataframe of dir, run, library, r1_uri, r2_uri
}
\description{
Generate a table from MPIBT SRA file structure
}
