% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_utils.R
\name{fqinfo}
\alias{fqinfo}
\title{Get basic statistics on a FastQ file}
\usage{
fqinfo(fastqgz, compress_ratio = 5.2173)
}
\arguments{
\item{fastqgz}{The path to a fastq.gz file to analyse}

\item{compress_ratio}{The average compression ratio for your data (bytes uncompressed/bytes compressed). You can estimate this manually for higher accuracy, however the default is reasonably consistent and any inaccuracy will be simply off by some scaling factor.}
}
\value{
A tibble() of read_length, estimate_n_reads, file_size
}
\description{
Get basic statistics on a FastQ file
}
