% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kview.R
\name{kview}
\alias{kview}
\title{View a dataframe with LibreOffice}
\usage{
kview(df)
}
\arguments{
\item{df}{Any dataframe writable with readr}
}
\value{
The input dataframe
}
\description{
View a dataframe with LibreOffice
}
