% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_delims.R
\name{read_delims}
\alias{read_delims}
\title{readr::read_delim() for many files at once}
\usage{
read_delims(files, ...)
}
\arguments{
\item{files}{A character vector of filenames}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  The resulting concatenated dataframe
}\if{html}{\out{</div>}}
}
\description{
readr::read_delim() for many files at once
}
