% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nawarn.R
\name{truey_falsey}
\alias{truey_falsey}
\title{Fuzzily interpret booleans as often written by hand on field notes}
\usage{
truey_falsey(x)
}
\arguments{
\item{x}{character representing boolean values}
}
\value{
The "after" input
}
\description{
Fuzzily interpret booleans as often written by hand on field notes
}
