% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nawarn.R
\name{whona}
\alias{whona}
\title{Check for coerced values, printing the values that were coerced to NA.}
\usage{
whona(x, y)
}
\arguments{
\item{x}{before conversion}

\item{y}{after conversion}
}
\value{
The "after" input
}
\description{
Check for coerced values, printing the values that were coerced to NA.
}
